requirements.txt
fastapi
uvicorn[standard]
celery
redis
python-dotenv
pymupdf
google-genai
edge-tts
pydub
# boto3
pymongo
httpx
python-multipart
supabase